requirements.txt
asgiref==3.4.1
Django==4.0.6
djangorestframework==3.13.1
django-fullurl==1.1
mysqlclient==2.0.3
Pillow==8.1.2
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
